+++ src/frequenz/sdk/microgrid/_power_distributing/_component_managers/__init__.py
@@ -7,10 +7,12 @@ from ._battery_manager import BatteryManager
 from ._component_manager import ComponentManager
 from ._ev_charger_manager import EVChargerManager
 from ._pv_inverter_manager import PVManager
+from ._steam_boiler_manager import SteamBoilerManager
 
 __all__ = [
     "BatteryManager",
     "ComponentManager",
     "EVChargerManager",
     "PVManager",
+    "SteamBoilerManager",
 ]