+++ tests/timeseries/_steam_boiler_pool/__init__.py
@@ -1,0 +1,4 @@
+# License: MIT
+# Copyright © 2026 Frequenz Energy-as-a-Service GmbH
+
+"""Test the steam boiler pool control methods."""